requirements.txt
cityhash==0.4.7
pycryptodome==3.19.0
tabulate==0.9.0
